fix(wallet-setup): navigation
