+++ apps/wallet-mobile/src/features/Send/common/navigation.ts
@@ -21,7 +21,7 @@ export const useNavigateTo = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [
                 {name: 'wallet-selection'},
