+++ apps/wallet-mobile/src/features/Swap/common/navigation.ts
@@ -20,7 +20,7 @@ export const useNavigateTo = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [
                 {name: 'wallet-selection'},
