+++ apps/wallet-mobile/src/navigation.tsx
@@ -179,7 +179,7 @@ export type SwapTokenRoutes = {
   'swap-select-pool': undefined
   'swap-submitted-tx': {txId: string}
   'swap-failed-tx': undefined
-  'app-root': undefined
+  'manage-wallets': undefined
 }
 export type SwapTokenRouteseNavigation = StackNavigationProp<SwapTokenRoutes>
 
@@ -201,7 +201,7 @@ export type ExchangeRoutes = {
   'exchange-result': undefined
   'exchange-select-buy-provider': undefined
   'exchange-select-sell-provider': undefined
-  'app-root': undefined
+  'manage-wallets': undefined
 }
 
 export type ExchangeRoutesNavigation = StackNavigationProp<ExchangeRoutes>
@@ -305,13 +305,15 @@ export type AppRoutes = {
   'first-run': NavigatorScreenParams<FirstRunRoutes>
   developer: undefined
   storybook: undefined
-  'app-root': NavigatorScreenParams<WalletStackRoutes>
+  'manage-wallets': NavigatorScreenParams<WalletStackRoutes>
   'custom-pin-auth': undefined
   'exchange-result': undefined
   'bio-auth-initial': undefined
   'enable-login-with-pin': undefined
   'agreement-changed-notice': undefined
   modal: undefined
+  'choose-biometric-login': undefined
+  'setup-wallet': undefined
 }
 export type AppRouteNavigation = StackNavigationProp<AppRoutes>
 
@@ -347,7 +349,7 @@ export const useWalletNavigation = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [
                 {name: 'wallet-selection'},
@@ -376,7 +378,7 @@ export const useWalletNavigation = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [
                 {name: 'wallet-selection'},
@@ -401,7 +403,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToStartTransfer: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'main-wallet-routes',
         params: {
           screen: 'history',
@@ -417,7 +419,7 @@ export const useWalletNavigation = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [
                 {
@@ -438,7 +440,7 @@ export const useWalletNavigation = () => {
         index: 0,
         routes: [
           {
-            name: 'app-root',
+            name: 'manage-wallets',
             state: {
               routes: [{name: 'wallet-selection'}],
             },
@@ -448,7 +450,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToStakingDashboard: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'main-wallet-routes',
         params: {
           screen: 'staking-dashboard',
@@ -460,7 +462,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToSettings: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'settings',
         params: {
           screen: 'main-settings',
@@ -469,7 +471,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToTxHistory: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'main-wallet-routes',
         params: {
           screen: 'history',
@@ -481,7 +483,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToNftGallery: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'main-wallet-routes',
         params: {
           screen: 'nfts',
@@ -493,7 +495,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToAppSettings: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'settings',
         params: {
           screen: 'app-settings',
@@ -502,7 +504,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToCollateralSettings: (params?: SettingsStackRoutes['manage-collateral']) => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'settings',
         params: {
           screen: 'manage-collateral',
@@ -512,7 +514,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToAnalyticsSettings: () => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'toggle-analytics-settings',
         params: {
           screen: 'settings',
@@ -521,7 +523,7 @@ export const useWalletNavigation = () => {
     },
 
     navigateToGovernanceCentre: ({navigateToStakingOnSuccess = false} = {}) => {
-      navigation.navigate('app-root', {
+      navigation.navigate('manage-wallets', {
         screen: 'governance',
         params: {
           screen: 'staking-gov-home',
