+++ apps/wallet-mobile/translations/messages/src/AppNavigator.json
@@ -4,14 +4,14 @@
     "defaultMessage": "!!!Enter PIN",
     "file": "src/AppNavigator.tsx",
     "start": {
-      "line": 186,
+      "line": 219,
       "column": 17,
-      "index": 6326
+      "index": 7782
     },
     "end": {
-      "line": 189,
+      "line": 222,
       "column": 3,
-      "index": 6416
+      "index": 7872
     }
   },
   {
@@ -19,14 +19,14 @@
     "defaultMessage": "!!!Set PIN",
     "file": "src/AppNavigator.tsx",
     "start": {
-      "line": 190,
+      "line": 223,
       "column": 18,
-      "index": 6436
+      "index": 7892
     },
     "end": {
-      "line": 193,
+      "line": 226,
       "column": 3,
-      "index": 6534
+      "index": 7990
     }
   },
   {
@@ -34,14 +34,14 @@
     "defaultMessage": "!!!Auth with OS changes",
     "file": "src/AppNavigator.tsx",
     "start": {
-      "line": 194,
+      "line": 227,
       "column": 25,
-      "index": 6561
+      "index": 8017
     },
     "end": {
-      "line": 197,
+      "line": 230,
       "column": 3,
-      "index": 6675
+      "index": 8131
     }
   },
   {
@@ -49,14 +49,14 @@
     "defaultMessage": "!!!Auth with OS changed detected",
     "file": "src/AppNavigator.tsx",
     "start": {
-      "line": 198,
+      "line": 231,
       "column": 27,
-      "index": 6704
+      "index": 8160
     },
     "end": {
-      "line": 201,
+      "line": 234,
       "column": 3,
-      "index": 6825
+      "index": 8281
     }
   }
 ]
