+++ apps/wallet-mobile/translations/messages/src/Dashboard/Dashboard.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 245,
       "column": 23,
-      "index": 7836
+      "index": 7842
     },
     "end": {
       "line": 248,
       "column": 3,
-      "index": 7969
+      "index": 7975
     }
   }
 ]
